Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-kw-import-key-unwrap-raw-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test unwrapping a raw key with AES-KW using an imported key");

jsTestIsAsync = true;

var extractable = true;
var rawKeyASCII = "jnOw99oOZFLIEPMr";
var rawKey = asciiToUint8Array(rawKeyASCII);
var wrappedKey = hexStringToUint8Array("d64787ab3e048dbdc30bb62781c9f18e58ad7dbfc64aab16");

crypto.subtle.importKey("raw", rawKey, "aes-kw", extractable, ["wrapKey", "unwrapKey"]).then(function(unwrappingKey) {
    return crypto.subtle.unwrapKey("raw", wrappedKey, unwrappingKey, "AES-KW", "AES-CBC", extractable, ["encrypt", "decrypt"]);
}).then(function(cryptoKey) {
    return crypto.subtle.exportKey("raw", cryptoKey);
}).then(function(result) {
    unwrappedKey = result;

    shouldBe("bytesToASCIIString(unwrappedKey)", "rawKeyASCII");

    finishJSTest();
});

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
